+++ tinder-client/src/main/java/ru/prerev/tinderclient/telegrambot/Bot.java
@@ -32,15 +32,15 @@ public final class Bot extends TelegramLongPollingBot {
         if (update.hasMessage() && update.getMessage().hasText()) {
             Long chatId = update.getMessage().getChatId();
             String message_text = update.getMessage().getText();
-            authorizeUser(chatId, message_text);
+            authorizeUser(chatId.intValue(), message_text);
         } else if (update.hasCallbackQuery()) {
             Long chatId = update.getCallbackQuery().getMessage().getChatId();
             String message_text = update.getCallbackQuery().getData();
-            authorizeUser(chatId, message_text);
+            authorizeUser(chatId.intValue(), message_text);
         }
     }
 
-    private void authorizeUser(Long chatId, String message_text) {
+    private void authorizeUser(Integer chatId, String message_text) {
         authorizer.setBot(this);
         try {
             authorizer.authorize(chatId, message_text);
